chore: implement review comments
