+++ apps/backend/src/modules/deployment/deployment.app.test.ts
@@ -84,7 +84,7 @@ import { RegistrationDomain } from '../registration/registration.domain';
 import { ServiceInstanceDomain } from '../service/instance/service-instance.domain';
 import { CompetitorDomain } from './competitor/competitor.domain';
 import {
-  BUNDLE_ACTIVATION_CANCELLATION_REASON,
+  BUNDLE_REQUEST_CANCELLATION_REASON,
   DeploymentApp,
   XTM_PLATFORM_BUNDLE_SERVICE_INSTANCE_NAME,
 } from './deployment.app';
@@ -232,6 +232,7 @@ describe('deployment app', () => {
       );
     });
     it('should throw when service definition is not found', async () => {
+      // When
       const call = DeploymentApp.createDeploymentRequest({
         ...TEST_DEPLOYMENT,
         products: ['unknown-platform' as PlatformIdentifier],
@@ -243,20 +244,24 @@ describe('deployment app', () => {
         ],
       });
 
+      // Then
       await expect(call).rejects.toThrow(ErrorCode.ServiceDefinitionNotFound);
     });
     it('should throw when a requested product has no use case', async () => {
+      // When
       const call = DeploymentApp.createDeploymentRequest({
         ...TEST_DEPLOYMENT,
         use_cases_by_product: [],
       });
 
+      // Then
       await expect(call).rejects.toThrow(
         BadRequestErrorCode.InvalidUseCasesForProducts
       );
     });
 
     it('should throw when a use case targets a product that was not requested', async () => {
+      // When
       const call = DeploymentApp.createDeploymentRequest({
         ...TEST_DEPLOYMENT,
         type: DeploymentRequestDeploymentType.Bundle,
@@ -273,11 +278,13 @@ describe('deployment app', () => {
         ],
       });
 
+      // Then
       await expect(call).rejects.toThrow(
         BadRequestErrorCode.InvalidUseCasesForProducts
       );
     });
     it('should throw when two use cases target the same product', async () => {
+      // When
       const call = DeploymentApp.createDeploymentRequest({
         ...TEST_DEPLOYMENT,
         type: DeploymentRequestDeploymentType.Bundle,
@@ -294,6 +301,7 @@ describe('deployment app', () => {
         ],
       });
 
+      // Then
       await expect(call).rejects.toThrow(
         BadRequestErrorCode.InvalidUseCasesForProducts
       );
@@ -305,11 +313,13 @@ describe('deployment app', () => {
     ])(
       'should throw InvalidProductsForDeploymentType for a trial with products %s',
       async (products) => {
+        // When
         const call = DeploymentApp.createDeploymentRequest({
           ...TEST_DEPLOYMENT,
           products,
         });
 
+        // Then
         await expect(call).rejects.toThrow(
           BadRequestErrorCode.InvalidProductsForDeploymentType
         );
@@ -323,12 +333,14 @@ describe('deployment app', () => {
     ])(
       'should throw InvalidProductsForDeploymentType for a bundle with products %s',
       async (products) => {
+        // When
         const call = DeploymentApp.createDeploymentRequest({
           ...TEST_DEPLOYMENT,
           type: DeploymentRequestDeploymentType.Bundle,
           products,
         });
 
+        // Then
         await expect(call).rejects.toThrow(
           BadRequestErrorCode.InvalidProductsForDeploymentType
         );
@@ -336,6 +348,7 @@ describe('deployment app', () => {
     );
 
     it("should throw when a bundle product's service definition is not found", async () => {
+      // When
       const call = DeploymentApp.createDeploymentRequest({
         ...TEST_DEPLOYMENT,
         type: DeploymentRequestDeploymentType.Bundle,
@@ -351,6 +364,7 @@ describe('deployment app', () => {
         ],
       });
 
+      // Then
       await expect(call).rejects.toThrow(ErrorCode.ServiceDefinitionNotFound);
     });
 
@@ -360,9 +374,11 @@ describe('deployment app', () => {
         vi.spyOn(DeploymentQuotaDomain, 'reservePlace').mockResolvedValue({
           isPlaceAvailable: true,
         });
+        vi.spyOn(DeploymentQuotaDomain, 'freePlace').mockResolvedValue();
       });
 
       it('should create a bundle deployment request with a child trial deployment request per product', async () => {
+        // When
         const bundle = await DeploymentApp.createDeploymentRequest({
           ...TEST_DEPLOYMENT,
           type: DeploymentRequestDeploymentType.Bundle,
@@ -383,6 +399,7 @@ describe('deployment app', () => {
           ],
         });
 
+        // Then
         expect(bundle).toMatchObject({
           type: DeploymentRequestDeploymentType.Bundle,
           platform_identifier: null,
@@ -471,6 +488,7 @@ describe('deployment app', () => {
       });
 
       it('should bypass the free trial limit check for bundle products', async () => {
+        // Given
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {
             organization_requester_id:
@@ -481,6 +499,7 @@ describe('deployment app', () => {
           }
         );
 
+        // When
         const bundle = await DeploymentApp.createDeploymentRequest({
           ...TEST_DEPLOYMENT,
           type: DeploymentRequestDeploymentType.Bundle,
@@ -493,6 +512,7 @@ describe('deployment app', () => {
           ],
         });
 
+        // Then
         expect(bundle.id).toBeDefined();
         const children = await TestHelper.deploymentRequest.loadMany({
           parent_id: bundle.id as DeploymentRequestId,
@@ -501,6 +521,7 @@ describe('deployment app', () => {
       });
 
       it('should roll back the whole bundle when a child fails to be created', async () => {
+        // Given
         const originalRegisterNewPlatform =
           RegistrationDomain.registerNewPlatform;
         let callCount = 0;
@@ -514,6 +535,7 @@ describe('deployment app', () => {
             return originalRegisterNewPlatform(args);
           });
 
+        // When
         const call = DeploymentApp.createDeploymentRequest({
           ...TEST_DEPLOYMENT,
           type: DeploymentRequestDeploymentType.Bundle,
@@ -534,6 +556,7 @@ describe('deployment app', () => {
           ],
         });
 
+        // Then
         await expect(call).rejects.toThrow('boom');
         expect(registerNewPlatformSpy).toHaveBeenCalledTimes(2);
 
@@ -547,6 +570,142 @@ describe('deployment app', () => {
         });
         expect(serviceInstances).toBeUndefined();
       });
+
+      describe('standalone trials cancellation', () => {
+        it('should cancel ongoing standalone trials of the organization when the bundle is requested', async () => {
+          // Given an ongoing standalone trial of the requesting organization
+          const standalone =
+            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
+              {
+                organization_requester_id:
+                  TEST_ORGANIZATIONS.SECOND_ORGANIZATION.ID,
+                hub_status: DeploymentRequestHubStatus.Active,
+              }
+            );
+
+          // When the bundle is requested
+          await DeploymentApp.createDeploymentRequest({
+            ...TEST_DEPLOYMENT,
+            type: DeploymentRequestDeploymentType.Bundle,
+            products: [PlatformIdentifier.Xtmone, PlatformIdentifier.Opencti],
+          });
+
+          // Then the standalone trial is cancelled
+          await TestHelper.deploymentRequest.assertProperties(standalone.id, {
+            hub_status: DeploymentRequestHubStatus.Cancelled,
+            cancellation_reason: BUNDLE_REQUEST_CANCELLATION_REASON,
+          });
+        });
+
+        it('should delete the Auth0 audience of cancelled standalone trials', async () => {
+          // Given an ongoing standalone trial of the requesting organization
+          const deleteAudienceSpy = vi.spyOn(
+            auth0ClientMock,
+            'deleteAudienceAPI'
+          );
+
+          const standalone =
+            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
+              {
+                organization_requester_id:
+                  TEST_ORGANIZATIONS.SECOND_ORGANIZATION.ID,
+                hub_status: DeploymentRequestHubStatus.Active,
+                platform_id: uuidv4(),
+              }
+            );
+
+          // When the bundle is requested
+          await DeploymentApp.createDeploymentRequest({
+            ...TEST_DEPLOYMENT,
+            type: DeploymentRequestDeploymentType.Bundle,
+            products: [PlatformIdentifier.Xtmone, PlatformIdentifier.Opencti],
+          });
+
+          // Then the Auth0 audience of the cancelled standalone trial is deleted
+          expect(deleteAudienceSpy).toHaveBeenCalledWith(
+            standalone.organization_requester_id,
+            standalone.platform_id
+          );
+        });
+
+        it('should not cancel standalone trials of another organization', async () => {
+          // Given an ongoing standalone trial owned by another organization
+          const standalone =
+            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
+              {
+                organization_requester_id: TEST_ORGANIZATIONS.FILIGRAN.ID,
+                hub_status: DeploymentRequestHubStatus.Active,
+              }
+            );
+
+          // When a bundle is requested by another organization
+          await DeploymentApp.createDeploymentRequest({
+            ...TEST_DEPLOYMENT,
+            type: DeploymentRequestDeploymentType.Bundle,
+            products: [PlatformIdentifier.Xtmone, PlatformIdentifier.Opencti],
+          });
+
+          const untouched =
+            await DeploymentRequestDomain.loadDeploymentRequestBy({
+              id: standalone.id,
+            });
+
+          // Then the other organization's standalone trial is untouched
+          expect(untouched).toMatchObject({
+            hub_status: DeploymentRequestHubStatus.Active,
+          });
+        });
+
+        it('should not cancel the children of the bundle itself', async () => {
+          // When the bundle is requested with its child trials
+          const bundle = await DeploymentApp.createDeploymentRequest({
+            ...TEST_DEPLOYMENT,
+            type: DeploymentRequestDeploymentType.Bundle,
+            products: [PlatformIdentifier.Xtmone, PlatformIdentifier.Opencti],
+          });
+
+          const children = await TestHelper.deploymentRequest.loadMany({
+            parent_id: bundle.id as DeploymentRequestId,
+          });
+
+          // Then the bundle's own children are untouched
+          expect(children).not.toHaveLength(0);
+          children.forEach((child) => {
+            expect(child.hub_status).not.toBe(
+              DeploymentRequestHubStatus.Cancelled
+            );
+          });
+        });
+
+        it('should not retroactively cancel standalone trials created after the bundle was requested', async () => {
+          // Given a bundle already requested for the organization
+          await DeploymentApp.createDeploymentRequest({
+            ...TEST_DEPLOYMENT,
+            type: DeploymentRequestDeploymentType.Bundle,
+            products: [PlatformIdentifier.Xtmone, PlatformIdentifier.Opencti],
+          });
+
+          // When an ongoing standalone trial is created after the bundle request
+          const standalone =
+            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
+              {
+                organization_requester_id:
+                  TEST_ORGANIZATIONS.SECOND_ORGANIZATION.ID,
+                hub_status: DeploymentRequestHubStatus.Active,
+              }
+            );
+
+          const untouched =
+            await DeploymentRequestDomain.loadDeploymentRequestBy({
+              id: standalone.id,
+            });
+
+          // Then the standalone trial is untouched
+          expect(untouched).toMatchObject({
+            hub_status: DeploymentRequestHubStatus.Active,
+          });
+        });
+      });
     });
   });
   describe('domains blacklist', () => {
@@ -558,6 +717,7 @@ describe('deployment app', () => {
       // Given
       await TestHelper.competitor.create();
 
+      // When
       const call = DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
       // Then
@@ -573,6 +733,7 @@ describe('deployment app', () => {
         domain: 'not-blocked.com',
       });
 
+      // When
       const deployment =
         await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
@@ -581,8 +742,10 @@ describe('deployment app', () => {
     });
 
     it('should allow deployment when no competitors exist', async () => {
+      // Given
       requestContext.set(requestContextRegistererUserSecondOrga);
 
+      // When
       const deployment =
         await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
@@ -600,12 +763,14 @@ describe('deployment app', () => {
       `(
         'should send a telemetry event when trial for $product platform is launched',
         async ({ product, targetProduct, source, telemetrySource }) => {
+          // Given
           requestContext.set(requestContextRegistererUserSecondOrga);
 
           vi.useFakeTimers();
           const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
           vi.setSystemTime(date);
 
+          // When
           const deployment = await DeploymentApp.createDeploymentRequest({
             activity_sector:
               DeploymentRequestActivitySector.ComputerNetworkSecurity,
@@ -622,6 +787,7 @@ describe('deployment app', () => {
             source,
           });
 
+          // Then
           expect(telemetrySpy).toHaveBeenCalledExactlyOnceWith({
             '@timestamp': '2025-02-03T13:12:15.000Z',
             event_type: TelemetryEventType.CREATE_DEPLOYMENT,
@@ -647,6 +813,7 @@ describe('deployment app', () => {
         }
       );
       it('should not throw when an error is thrown by telemetry', async () => {
+        // Given
         requestContext.set(requestContextRegistererUserSecondOrga);
 
         vi.useFakeTimers();
@@ -654,9 +821,11 @@ describe('deployment app', () => {
         vi.setSystemTime(date);
         telemetrySpy.mockRejectedValue(new Error('UNKNOWN'));
 
+        // When
         const deployment =
           await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
+        // Then
         expect(deployment).toBeDefined();
       });
     });
@@ -664,9 +833,13 @@ describe('deployment app', () => {
     describe('mail', () => {
       describe('development environment', () => {
         it('should send a mail if status is pending to dev team', async () => {
+          // Given
           requestContext.set(requestContextAdminUser);
+
+          // When
           await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
+          // Then
           expect(mockSendMail).toHaveBeenCalledTimes(2);
 
           expect(mockSendMail).toHaveBeenNthCalledWith(1, {
@@ -696,13 +869,17 @@ describe('deployment app', () => {
         });
 
         it('should send a mail if there is no space available', async () => {
+          // Given
           requestContext.set(requestContextRegistererUserSecondOrga);
 
           vi.spyOn(DeploymentQuotaDomain, 'reservePlace').mockResolvedValue({
             isPlaceAvailable: false,
           });
+
+          // When
           await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
+          // Then
           expect(mockSendMail).toHaveBeenCalledTimes(2);
 
           expect(mockSendMail).toHaveBeenNthCalledWith(1, {
@@ -746,10 +923,13 @@ describe('deployment app', () => {
         });
 
         it('should send a mail if status is pending to dev team', async () => {
+          // Given
           requestContext.set(requestContextAdminUser);
 
+          // When
           await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
+          // Then
           expect(mockSendMail).toHaveBeenCalledTimes(2);
 
           expect(mockSendMail).toHaveBeenNthCalledWith(1, {
@@ -779,13 +959,17 @@ describe('deployment app', () => {
         });
 
         it('should send a mail if there is no space available', async () => {
+          // Given
           requestContext.set(requestContextRegistererUserSecondOrga);
 
           vi.spyOn(DeploymentQuotaDomain, 'reservePlace').mockResolvedValue({
             isPlaceAvailable: false,
           });
+
+          // When
           await DeploymentApp.createDeploymentRequest(TEST_DEPLOYMENT);
 
+          // Then
           expect(mockSendMail).toHaveBeenCalledTimes(2);
 
           expect(mockSendMail).toHaveBeenNthCalledWith(1, {
@@ -825,15 +1009,18 @@ describe('deployment app', () => {
     });
 
     it('should return created deployment requests', async () => {
+      // Given a created deployment request
       const deploymentRequest =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {}
         );
 
+      // When loading the platform deployment requests
       const deployments = await DeploymentApp.loadPlatformDeploymentRequests({
         first: 10,
       });
 
+      // Then the created deployment request is returned
       expect(deployments.totalCount).toBe('1');
       expect(deployments.edges[0]?.node).toStrictEqual({
         ...deploymentRequest,
@@ -852,6 +1039,7 @@ describe('deployment app', () => {
     });
 
     it('should return platform_url when PlatformConfiguration exists', async () => {
+      // Given a deployment request with an active platform configuration
       const deploymentRequest =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {}
@@ -863,6 +1051,7 @@ describe('deployment app', () => {
         status: PlatformConfigurationStatus.Active,
       });
 
+      // When loading the platform deployment requests
       const deployments = await DeploymentApp.loadPlatformDeploymentRequests({
         first: 10,
       });
@@ -871,6 +1060,7 @@ describe('deployment app', () => {
         (edge) => edge.node.id === deploymentRequest!.id
       );
 
+      // Then the platform_url from the configuration is returned
       expect(deployment?.node.platform_url).toBe(
         'https://test-platform.opencti.io'
       );
@@ -881,6 +1071,7 @@ describe('deployment app', () => {
     });
 
     it('should return out-of-sync deployment requests by default', async () => {
+      // Given one out-of-sync deployment request and one synced deployment request
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {
           hub_status: DeploymentRequestHubStatus.Pending,
@@ -896,10 +1087,12 @@ describe('deployment app', () => {
         }
       );
 
+      // When loading the platform deployment requests without filters
       const deployments = await DeploymentApp.loadPlatformDeploymentRequests({
         first: 10,
       });
 
+      // Then only the out-of-sync deployment request is returned
       expect(deployments.totalCount).toBe('1');
       expect(deployments.edges).toHaveLength(1);
       expect(deployments.edges[0]?.node?.hub_status).toBe(
@@ -908,6 +1101,7 @@ describe('deployment app', () => {
     });
 
     it('should return out-of-sync deployments even with other filters', async () => {
+      // Given one out-of-sync deployment request and one synced deployment request
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {
           target_state: DeploymentRequestPlatformState.Active,
@@ -922,6 +1116,7 @@ describe('deployment app', () => {
         }
       );
 
+      // When loading the platform deployment requests with an additional filter
       const deployments = await DeploymentApp.loadPlatformDeploymentRequests({
         first: 10,
         filters: [
@@ -932,6 +1127,7 @@ describe('deployment app', () => {
         ],
       });
 
+      // Then only the out-of-sync deployment request is returned
       expect(deployments.totalCount).toBe('1');
       expect(deployments.edges).toHaveLength(1);
       expect(deployments.edges[0]?.node?.hub_status).toBe(
@@ -940,6 +1136,7 @@ describe('deployment app', () => {
     });
 
     it('should filter multiple out-of-sync scenarios correctly', async () => {
+      // Given deployment requests covering synced and out-of-sync scenarios
       // Out-of-sync: NULL target vs NULL actual (both NULL = synced, should NOT appear)
       const synced1 =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
@@ -1000,14 +1197,15 @@ describe('deployment app', () => {
           }
         );
 
+      // When loading the platform deployment requests
       const deployments = await DeploymentApp.loadPlatformDeploymentRequests({
         first: 10,
       });
 
+      // Then only the out-of-sync deployment requests are returned
       expect(deployments.totalCount).toBe('3');
       expect(deployments.edges).toHaveLength(3);
 
-      // Verify only out-of-sync deployments are returned
       const returnedIds = deployments.edges.map((edge) => edge.node.id);
       expect(returnedIds).toContain(outOfSync1!.id);
       expect(returnedIds).toContain(outOfSync2!.id);
@@ -1018,6 +1216,7 @@ describe('deployment app', () => {
     });
 
     it('should return filtered deployment requests only', async () => {
+      // Given deployment requests that do not match every filter criterion at once
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {}
       );
@@ -1036,6 +1235,7 @@ describe('deployment app', () => {
         }
       );
 
+      // When loading the platform deployment requests with combined filters
       const deployments = await DeploymentApp.loadPlatformDeploymentRequests({
         first: 10,
         filters: [
@@ -1054,6 +1254,7 @@ describe('deployment app', () => {
         ],
       });
 
+      // Then no deployment request matches all filters
       expect(deployments.totalCount).toBe('0');
       expect(deployments.edges).toHaveLength(0);
     });
@@ -1074,6 +1275,7 @@ describe('deployment app', () => {
     });
 
     it('should update a deployment request', async () => {
+      // When updating the deployment request
       const deployment = await DeploymentApp.updateDeploymentRequest({
         id: initialDeployment?.id as DeploymentRequestId,
         actual_state: DeploymentRequestPlatformState.Active,
@@ -1098,6 +1300,8 @@ describe('deployment app', () => {
       const subscription = await SubscriptionDomain.loadSubscriptionBy({
         service_instance_id: dbDeploymentRequest.service_instance_id,
       });
+
+      // Then the deployment request is updated with the new values
       expect(dbDeploymentRequest).toMatchObject({
         activity_sector:
           DeploymentRequestActivitySector.ComputerNetworkSecurity,
@@ -1137,6 +1341,7 @@ describe('deployment app', () => {
     });
 
     it('with Active status for OpenCTI, it should create OpenCTI ServiceGroups (Admin, Analyst, Reader) with admin user', async () => {
+      // When updating the deployment request to Active
       const deployment = await DeploymentApp.updateDeploymentRequest({
         id: initialDeployment?.id as DeploymentRequestId,
         actual_state: DeploymentRequestPlatformState.Active,
@@ -1152,6 +1357,8 @@ describe('deployment app', () => {
       const serviceGroups = await ServiceGroupDomain.loadServiceGroups({
         service_instance_id: dbDeploymentRequest!.service_instance_id,
       });
+
+      // Then the OpenCTI ServiceGroups are created
       expect(serviceGroups).toHaveLength(3);
       expect(serviceGroups.map((g) => g.name).sort()).toEqual([
         ServiceGroupName.Admin,
@@ -1187,6 +1394,7 @@ describe('deployment app', () => {
     });
 
     it('with Active status for OpenAEV, it should create OpenAEV ServiceGroups (Admin, Manager, Observer) with admin user', async () => {
+      // Given an OpenAEV deployment request
       const openaevDeployment =
         (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {
@@ -1197,6 +1405,7 @@ describe('deployment app', () => {
           }
         )) as DeploymentRequest;
 
+      // When updating the deployment request to Active
       const deployment = await DeploymentApp.updateDeploymentRequest({
         id: openaevDeployment.id as DeploymentRequestId,
         actual_state: DeploymentRequestPlatformState.Active,
@@ -1212,6 +1421,8 @@ describe('deployment app', () => {
       const serviceGroups = await ServiceGroupDomain.loadServiceGroups({
         service_instance_id: dbDeploymentRequest!.service_instance_id,
       });
+
+      // Then the OpenAEV ServiceGroups are created
       expect(serviceGroups).toHaveLength(3);
       expect(serviceGroups.map((g) => g.name).sort()).toEqual([
         ServiceGroupName.Admin,
@@ -1246,6 +1457,7 @@ describe('deployment app', () => {
     });
 
     it('with Active status for XTM One, it should not create any ServiceGroup', async () => {
+      // Given an XTM One deployment request
       const xtmoneDeployment =
         (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {
@@ -1256,6 +1468,7 @@ describe('deployment app', () => {
           }
         )) as DeploymentRequest;
 
+      // When updating the deployment request to Active
       const deployment = await DeploymentApp.updateDeploymentRequest({
         id: xtmoneDeployment.id as DeploymentRequestId,
         actual_state: DeploymentRequestPlatformState.Active,
@@ -1271,16 +1484,20 @@ describe('deployment app', () => {
       const serviceGroups = await ServiceGroupDomain.loadServiceGroups({
         service_instance_id: dbDeploymentRequest!.service_instance_id,
       });
+
+      // Then no ServiceGroup is created
       expect(serviceGroups).toHaveLength(0);
     });
 
     it('should set platform registration status to inactive when actual state is removed', async () => {
+      // Given an active platform configuration for the deployment request
       await TestHelper.platformConfiguration.create({
         service_instance_id: initialDeployment.service_instance_id,
         status: PlatformConfigurationStatus.Active,
       });
 
       try {
+        // When updating the deployment request to Removed
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Removed,
@@ -1291,6 +1508,7 @@ describe('deployment app', () => {
             service_instance_id: initialDeployment.service_instance_id,
           });
 
+        // Then the platform configuration status is set to inactive
         expect(platformConfiguration?.status).toBe(
           PlatformConfigurationStatus.Inactive
         );
@@ -1302,10 +1520,13 @@ describe('deployment app', () => {
     });
 
     it('should throw if deployment request does not exist', async () => {
+      // Given a non-existent deployment request id
       const call = DeploymentApp.updateDeploymentRequest({
         id: uuidv4() as DeploymentRequestId,
         actual_state: DeploymentRequestPlatformState.Active,
       });
+
+      // When / Then updating the deployment request throws
       await expect(call).rejects.toThrow(
         NotFoundErrorCode.DeploymentRequestNotFound
       );
@@ -1329,6 +1550,7 @@ describe('deployment app', () => {
     ])(
       'should throw if status active and $description',
       async ({ start_date, end_date }) => {
+        // Given an active state update with missing start or end date
         const call = DeploymentApp.updateDeploymentRequest({
           id: initialDeployment.id,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1336,6 +1558,7 @@ describe('deployment app', () => {
           end_date,
         });
 
+        // When / Then updating the deployment request throws
         await expect(call).rejects.toThrow(
           BadRequestErrorCode.MissingStartOrEndDate
         );
@@ -1343,6 +1566,7 @@ describe('deployment app', () => {
     );
     describe('telemetry', () => {
       it('should send a telemetry event', async () => {
+        // Given a fixed system time and target dates for the update
         vi.useFakeTimers();
         const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
         vi.setSystemTime(date);
@@ -1350,6 +1574,7 @@ describe('deployment app', () => {
         const start_date = new Date(2025, 1, 3);
         const end_date = new Date(2025, 2, 3);
 
+        // When updating the deployment request to Active
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1359,6 +1584,7 @@ describe('deployment app', () => {
           failure_reason: 'not failed',
         });
 
+        // Then a telemetry event is sent with the update details
         expect(telemetrySpy).toHaveBeenCalledExactlyOnceWith({
           '@timestamp': '2025-02-03T13:12:15.000Z',
           event_type: TelemetryEventType.UPDATE_DEPLOYMENT,
@@ -1378,6 +1604,7 @@ describe('deployment app', () => {
       });
 
       it('should not send a telemetry event when data did not change', async () => {
+        // Given a first update to the deployment request
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Provisioning,
@@ -1385,15 +1612,18 @@ describe('deployment app', () => {
 
         telemetrySpy.mockClear();
 
+        // When updating the deployment request again with the same data
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Provisioning,
         });
 
+        // Then no telemetry event is sent
         expect(telemetrySpy).not.toHaveBeenCalled();
       });
 
       it('should not throw when telemetry throws an error', async () => {
+        // Given telemetry rejecting and a fixed system time
         vi.useFakeTimers();
         const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
         telemetrySpy.mockRejectedValue(new Error('UNKNOWN'));
@@ -1402,6 +1632,7 @@ describe('deployment app', () => {
         const start_date = new Date(2025, 1, 3);
         const end_date = new Date(2025, 2, 3);
 
+        // When updating the deployment request to Active
         const deployment = await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1410,15 +1641,20 @@ describe('deployment app', () => {
           platform_id: 'fake product instance id',
           failure_reason: 'not failed',
         });
+
+        // Then the update still succeeds
         expect(deployment).toBeDefined();
       });
     });
     describe('mail', () => {
       it('should send a mail in case deployment request is in provisioning (only first time)', async () => {
+        // When updating the deployment request to Provisioning for the first time
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Provisioning,
         });
+
+        // Then a mail is sent
         expect(mockSendMail).toHaveBeenCalledWith({
           to: TEST_ORGANIZATIONS.FILIGRAN.USERS.BYPASS.EMAIL,
           template: 'free_trial_provisioning',
@@ -1430,15 +1666,18 @@ describe('deployment app', () => {
 
         mockSendMail.mockClear();
 
+        // When updating the deployment request to Provisioning again
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Provisioning,
         });
 
+        // Then no mail is sent again
         expect(mockSendMail).not.toHaveBeenCalled();
       });
 
       it('should send a mail in case deployment request is in active (only first time)', async () => {
+        // Given an active platform configuration for the deployment
         vi.spyOn(
           PlatformConfigurationDomain,
           'loadConfigurationByPlatform'
@@ -1456,6 +1695,8 @@ describe('deployment app', () => {
           status: PlatformConfigurationStatus.Active,
           last_connectivity_check: new Date(),
         });
+
+        // When updating the deployment request to Active for the first time
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           start_date: new Date(2025, 12, 1),
@@ -1464,6 +1705,7 @@ describe('deployment app', () => {
           platform_id: uuidv4(),
         });
 
+        // Then a mail is sent
         expect(mockSendMail).toHaveBeenCalledWith({
           to: TEST_ORGANIZATIONS.FILIGRAN.USERS.BYPASS.EMAIL,
           template: 'free_trial_registered',
@@ -1480,18 +1722,22 @@ describe('deployment app', () => {
 
         mockSendMail.mockClear();
 
+        // When updating the deployment request to Active again
         await DeploymentApp.updateDeploymentRequest({
           id: initialDeployment?.id as DeploymentRequestId,
           start_date: new Date(2025, 12, 1),
           end_date: new Date(2026, 1, 1),
           actual_state: DeploymentRequestPlatformState.Active,
         });
+
+        // Then no mail is sent again
         expect(mockSendMail).not.toHaveBeenCalled();
       });
     });
 
     describe('bundle updates', () => {
       it('should update actual_state and platform_id only, ignoring bundle-unsupported fields', async () => {
+        // Given a pending bundle
         const bundle =
           await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -1506,6 +1752,7 @@ describe('deployment app', () => {
             }
           );
 
+        // When updating the bundle with both supported and bundle-unsupported fields
         const updated = await DeploymentApp.updateDeploymentRequest({
           id: bundle.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1516,6 +1763,7 @@ describe('deployment app', () => {
           url: 'https://should-be-ignored.example.com',
         });
 
+        // Then only actual_state and platform_id are applied
         expect(updated).toMatchObject({
           actual_state: DeploymentRequestPlatformState.Active,
           platform_id: 'bundle-platform-id',
@@ -1527,6 +1775,7 @@ describe('deployment app', () => {
         });
       });
       it('should not initialise a service group for a bundle', async () => {
+        // Given a pending bundle
         const bundle =
           await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -1537,6 +1786,7 @@ describe('deployment app', () => {
             }
           );
 
+        // When updating the bundle to Active
         await DeploymentApp.updateDeploymentRequest({
           id: bundle.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1546,10 +1796,13 @@ describe('deployment app', () => {
         const serviceGroups = await ServiceGroupDomain.loadServiceGroups({
           service_instance_id: bundle.service_instance_id,
         });
+
+        // Then no service group is created for the bundle
         expect(serviceGroups).toHaveLength(0);
       });
 
       it('should transition hub_status through the same states as a product, driven by its own actual_state', async () => {
+        // Given a pending bundle
         const bundle =
           await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -1560,197 +1813,25 @@ describe('deployment app', () => {
             }
           );
 
+        // When updating the bundle's actual_state to Provisioning
         const provisioning = await DeploymentApp.updateDeploymentRequest({
           id: bundle.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Provisioning,
         });
+
+        // Then the bundle's hub_status becomes Provisioning
         expect(provisioning.hub_status).toBe(
           DeploymentRequestHubStatus.Provisioning
         );
 
+        // When updating the bundle's actual_state to Active
         const active = await DeploymentApp.updateDeploymentRequest({
           id: bundle.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
         });
+
+        // Then the bundle's hub_status becomes Active
         expect(active.hub_status).toBe(DeploymentRequestHubStatus.Active);
-      });
-      describe('standalone trials cancellation', () => {
-        beforeEach(() => {
-          vi.spyOn(DeploymentQuotaDomain, 'freePlace').mockResolvedValue();
-        });
-
-        it('should cancel ongoing standalone trials of the organization when the bundle becomes active', async () => {
-          const standalone =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                hub_status: DeploymentRequestHubStatus.Active,
-              }
-            );
-
-          const bundle =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                type: DeploymentRequestDeploymentType.Bundle,
-                platform_identifier: null,
-                hub_status: DeploymentRequestHubStatus.Pending,
-                actual_state: DeploymentRequestPlatformState.Unprovisioned,
-              }
-            );
-
-          await DeploymentApp.updateDeploymentRequest({
-            id: bundle.id as DeploymentRequestId,
-            actual_state: DeploymentRequestPlatformState.Active,
-          });
-
-          const cancelled =
-            await DeploymentRequestDomain.loadDeploymentRequestBy({
-              id: standalone.id,
-            });
-          expect(cancelled).toMatchObject({
-            hub_status: DeploymentRequestHubStatus.Cancelled,
-            cancellation_reason: BUNDLE_ACTIVATION_CANCELLATION_REASON,
-          });
-        });
-
-        it('should delete the Auth0 audience of cancelled standalone trials', async () => {
-          const deleteAudienceSpy = vi.spyOn(
-            auth0ClientMock,
-            'deleteAudienceAPI'
-          );
-
-          const standalone =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                hub_status: DeploymentRequestHubStatus.Active,
-                platform_id: uuidv4(),
-              }
-            );
-
-          const bundle =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                type: DeploymentRequestDeploymentType.Bundle,
-                platform_identifier: null,
-                hub_status: DeploymentRequestHubStatus.Pending,
-                actual_state: DeploymentRequestPlatformState.Unprovisioned,
-              }
-            );
-
-          await DeploymentApp.updateDeploymentRequest({
-            id: bundle.id as DeploymentRequestId,
-            actual_state: DeploymentRequestPlatformState.Active,
-          });
-
-          expect(deleteAudienceSpy).toHaveBeenCalledWith(
-            standalone.organization_requester_id,
-            standalone.platform_id
-          );
-        });
-
-        it('should not cancel standalone trials of another organization', async () => {
-          const standalone =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                organization_requester_id:
-                  TEST_ORGANIZATIONS.SECOND_ORGANIZATION.ID,
-                hub_status: DeploymentRequestHubStatus.Active,
-              }
-            );
-
-          const bundle =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                type: DeploymentRequestDeploymentType.Bundle,
-                platform_identifier: null,
-                hub_status: DeploymentRequestHubStatus.Pending,
-                actual_state: DeploymentRequestPlatformState.Unprovisioned,
-              }
-            );
-
-          await DeploymentApp.updateDeploymentRequest({
-            id: bundle.id as DeploymentRequestId,
-            actual_state: DeploymentRequestPlatformState.Active,
-          });
-
-          const untouched =
-            await DeploymentRequestDomain.loadDeploymentRequestBy({
-              id: standalone.id,
-            });
-          expect(untouched).toMatchObject({
-            hub_status: DeploymentRequestHubStatus.Active,
-          });
-        });
-
-        it('should not cancel the children of the bundle itself', async () => {
-          const bundle =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                type: DeploymentRequestDeploymentType.Bundle,
-                platform_identifier: null,
-                hub_status: DeploymentRequestHubStatus.Pending,
-                actual_state: DeploymentRequestPlatformState.Unprovisioned,
-              }
-            );
-
-          const child =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                parent_id: bundle.id as DeploymentRequestId,
-                hub_status: DeploymentRequestHubStatus.Active,
-              }
-            );
-
-          await DeploymentApp.updateDeploymentRequest({
-            id: bundle.id as DeploymentRequestId,
-            actual_state: DeploymentRequestPlatformState.Active,
-          });
-
-          const untouched =
-            await DeploymentRequestDomain.loadDeploymentRequestBy({
-              id: child.id,
-            });
-          expect(untouched).toMatchObject({
-            hub_status: DeploymentRequestHubStatus.Active,
-          });
-        });
-
-        it('should not cancel standalone trials again when an already active bundle is updated', async () => {
-          const bundle =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                type: DeploymentRequestDeploymentType.Bundle,
-                platform_identifier: null,
-                hub_status: DeploymentRequestHubStatus.Pending,
-                actual_state: DeploymentRequestPlatformState.Unprovisioned,
-              }
-            );
-
-          await DeploymentApp.updateDeploymentRequest({
-            id: bundle.id as DeploymentRequestId,
-            actual_state: DeploymentRequestPlatformState.Active,
-          });
-
-          const standalone =
-            await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
-              {
-                hub_status: DeploymentRequestHubStatus.Active,
-              }
-            );
-
-          await DeploymentApp.updateDeploymentRequest({
-            id: bundle.id as DeploymentRequestId,
-            actual_state: DeploymentRequestPlatformState.Active,
-            platform_id: 'bundle-platform-id',
-          });
-
-          const untouched =
-            await DeploymentRequestDomain.loadDeploymentRequestBy({
-              id: standalone.id,
-            });
-          expect(untouched).toMatchObject({
-            hub_status: DeploymentRequestHubStatus.Active,
-          });
-        });
       });
     });
 
@@ -1799,6 +1880,7 @@ describe('deployment app', () => {
       });
 
       it('should aggregate dates from children without changing the bundle hub_status', async () => {
+        // When updating childA to Active with its own dates
         await DeploymentApp.updateDeploymentRequest({
           id: childA.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1812,6 +1894,8 @@ describe('deployment app', () => {
           await DeploymentRequestDomain.loadDeploymentRequestBy({
             id: bundle.id as DeploymentRequestId,
           });
+
+        // Then the bundle aggregates the children dates without changing its hub_status
         expect(updatedBundle).toMatchObject({
           hub_status: DeploymentRequestHubStatus.Pending,
           start_date: new Date(2025, 1, 1),
@@ -1837,6 +1921,7 @@ describe('deployment app', () => {
       });
 
       it('should not update sibling children when the bundle itself is updated', async () => {
+        // When updating the bundle itself to Active
         await DeploymentApp.updateDeploymentRequest({
           id: bundle.id as DeploymentRequestId,
           actual_state: DeploymentRequestPlatformState.Active,
@@ -1851,6 +1936,8 @@ describe('deployment app', () => {
           await DeploymentRequestDomain.loadDeploymentRequestBy({
             id: childB.id as DeploymentRequestId,
           });
+
+        // Then the sibling children are left untouched
         expect(untouchedChildA?.hub_status).toBe(
           DeploymentRequestHubStatus.Provisioning
         );
@@ -1865,12 +1952,14 @@ describe('deployment app', () => {
       ])(
         'should not recompute dates for a %s bundle when a child is updated afterward',
         async (terminalHubStatus) => {
+          // Given a bundle in a terminal hub_status
           await DeploymentRequestDomain.updateDeploymentRequestById(
             bundle.id as DeploymentRequestId,
             { hub_status: terminalHubStatus }
           );
           telemetrySpy.mockClear();
 
+          // When updating a child of the bundle
           await DeploymentApp.updateDeploymentRequest({
             id: childA.id as DeploymentRequestId,
             actual_state: DeploymentRequestPlatformState.Active,
@@ -1883,6 +1972,8 @@ describe('deployment app', () => {
             await DeploymentRequestDomain.loadDeploymentRequestBy({
               id: bundle.id as DeploymentRequestId,
             });
+
+          // Then the bundle's terminal status and dates are not recomputed
           expect(updatedBundle?.hub_status).toBe(terminalHubStatus);
           expect(updatedBundle?.start_date).toBeNull();
           expect(updatedBundle?.end_date).toBeNull();
@@ -1899,10 +1990,12 @@ describe('deployment app', () => {
   });
   describe('loadTrialDeployments', () => {
     it('should return all available when no DeploymentRequest and no PlatformIdentifier specified', async () => {
+      // When loading trial deployments without a platform identifier filter
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.FILIGRAN.ID,
       });
 
+      // Then all platforms are reported as available
       expect(trialDeployments).toEqual({
         availableTrials: expect.arrayContaining([
           PlatformIdentifier.Opencti,
@@ -1914,11 +2007,13 @@ describe('deployment app', () => {
       expect(trialDeployments.availableTrials).toHaveLength(2);
     });
     it('should return only requested platform identifier specified as available when no DeploymentRequest exist', async () => {
+      // When loading trial deployments filtered to OpenCTI
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.FILIGRAN.ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then only OpenCTI is reported as available
       expect(trialDeployments).toEqual({
         availableTrials: [PlatformIdentifier.Opencti],
         deployed: [],
@@ -1927,6 +2022,7 @@ describe('deployment app', () => {
     });
 
     it('should return blacklisted = true if orga is blacklisted', async () => {
+      // Given an organization blacklisted as a competitor
       await CompetitorDomain.insertCompetitor({
         id: uuidv4() as CompetitorId,
         name: 'Filigran',
@@ -1934,11 +2030,13 @@ describe('deployment app', () => {
         domain: TEST_ORGANIZATIONS.FILIGRAN.DOMAINS.FIRST,
       });
 
+      // When loading trial deployments for that organization
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.FILIGRAN.ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then isBlacklisted is true
       expect(trialDeployments).toEqual({
         availableTrials: [PlatformIdentifier.Opencti],
         deployed: [],
@@ -1948,17 +2046,20 @@ describe('deployment app', () => {
       await TestHelper.competitor.delete({});
     });
     it('should return trial as available if the created one does not count in quota', async () => {
+      // Given a deployment request that does not count in the organization quota
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {
           counts_in_orga_quota: false,
         }
       );
 
+      // When loading trial deployments for that organization
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.FILIGRAN.ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then the platform is still reported as available
       expect(trialDeployments).toEqual({
         availableTrials: [PlatformIdentifier.Opencti],
         deployed: [],
@@ -1967,16 +2068,19 @@ describe('deployment app', () => {
     });
 
     it('should not return identifier as available when DeploymentRequest exist', async () => {
+      // Given an existing deployment request with a service instance and subscription
       const deploymentRequest =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {}
         );
 
+      // When loading trial deployments for that organization
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.FILIGRAN.ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then the platform is reported as deployed and no longer available
       expect(trialDeployments).toEqual({
         availableTrials: [],
         deployed: [
@@ -1989,16 +2093,19 @@ describe('deployment app', () => {
       });
     });
     it('should return data corresponding to the right organization', async () => {
+      // Given a deployment request for the first organization
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {}
       );
 
+      // When loading trial deployments for a different organization
       requestContext.set(requestContextAdminSecondOrga);
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.SECOND_ORGANIZATION.ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then the other organization's deployment request does not affect availability
       expect(trialDeployments).toEqual({
         availableTrials: [PlatformIdentifier.Opencti],
         deployed: [],
@@ -2006,6 +2113,7 @@ describe('deployment app', () => {
       });
     });
     it('should return not availablity and no deployed for personal space', async () => {
+      // Given a deployment request in the second organization and a user on their personal space
       requestContext.set(requestContextAdminSecondOrga);
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {}
@@ -2025,12 +2133,14 @@ describe('deployment app', () => {
         user: contextUserWithPersonalOrga.user,
       });
 
+      // When loading trial deployments for the personal space
       const trialDeployments = await DeploymentApp.loadTrialDeployments({
         organizationId:
           TEST_ORGANIZATIONS.SECOND_ORGANIZATION.USERS.SIMPLE.PERSONAL_SPACE_ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then no trial is available and nothing is deployed
       expect(trialDeployments).toEqual({
         availableTrials: [],
         deployed: [],
@@ -2038,25 +2148,31 @@ describe('deployment app', () => {
       });
     });
     it('should throw if user does not belong in the organization', async () => {
+      // Given a deployment request and a user from a different organization
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {}
       );
 
       requestContext.set(requestContextAdminSecondOrga);
+
+      // When loading trial deployments for an organization the user does not belong to
       const call = DeploymentApp.loadTrialDeployments({
         organizationId: TEST_ORGANIZATIONS.FILIGRAN.ID,
         platformIdentifiers: [PlatformIdentifier.Opencti],
       });
 
+      // Then it throws UserIsNotInOrganization
       await expect(call).rejects.toThrow(ErrorCode.UserIsNotInOrganization);
     });
   });
   describe('loadPlatformTrialStatus', () => {
     it('should return null hub_status and end_date when no bundle DeploymentRequest exists', async () => {
+      // When loading the platform trial status with no existing deployment request
       const platformTrialStatus = await DeploymentApp.loadPlatformTrialStatus(
         TEST_ORGANIZATIONS.FILIGRAN.ID
       );
 
+      // Then hub_status and end_date are null
       expect(platformTrialStatus).toEqual({
         ongoingStandaloneTrials: [],
         isBlacklisted: false,
@@ -2066,6 +2182,7 @@ describe('deployment app', () => {
     });
 
     it('should return the bundle hub_status and end_date when a bundle DeploymentRequest exists', async () => {
+      // Given an active bundle deployment request with an end date
       const endDate = new Date(Date.UTC(2025, 5, 1));
       const bundle =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
@@ -2077,10 +2194,12 @@ describe('deployment app', () => {
           }
         );
 
+      // When loading the platform trial status
       const platformTrialStatus = await DeploymentApp.loadPlatformTrialStatus(
         TEST_ORGANIZATIONS.FILIGRAN.ID
       );
 
+      // Then the bundle's hub_status and end_date are returned
       expect(platformTrialStatus).toEqual({
         ongoingStandaloneTrials: [],
         isBlacklisted: false,
@@ -2090,6 +2209,7 @@ describe('deployment app', () => {
     });
 
     it('should treat a bundle DeploymentRequest as non-existent when counts_in_orga_quota is false', async () => {
+      // Given an active bundle deployment request that does not count in the organization quota
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {
           type: DeploymentRequestDeploymentType.Bundle,
@@ -2100,10 +2220,12 @@ describe('deployment app', () => {
         }
       );
 
+      // When loading the platform trial status
       const platformTrialStatus = await DeploymentApp.loadPlatformTrialStatus(
         TEST_ORGANIZATIONS.FILIGRAN.ID
       );
 
+      // Then hub_status and end_date are null, as if the bundle did not exist
       expect(platformTrialStatus).toEqual({
         ongoingStandaloneTrials: [],
         isBlacklisted: false,
@@ -2113,6 +2235,7 @@ describe('deployment app', () => {
     });
 
     it('should return blacklisted = true if orga is blacklisted', async () => {
+      // Given an organization blacklisted as a competitor
       await CompetitorDomain.insertCompetitor({
         id: uuidv4() as CompetitorId,
         name: 'Filigran',
@@ -2120,10 +2243,12 @@ describe('deployment app', () => {
         domain: TEST_ORGANIZATIONS.FILIGRAN.DOMAINS.FIRST,
       });
 
+      // When loading the platform trial status
       const platformTrialStatus = await DeploymentApp.loadPlatformTrialStatus(
         TEST_ORGANIZATIONS.FILIGRAN.ID
       );
 
+      // Then isBlacklisted is true
       expect(platformTrialStatus).toEqual({
         ongoingStandaloneTrials: [],
         isBlacklisted: true,
@@ -2135,6 +2260,7 @@ describe('deployment app', () => {
     });
 
     it('should return null hub_status and end_date for personal space, even with a bundle in another organization', async () => {
+      // Given a bundle deployment request in the second organization and a user on their personal space
       requestContext.set(requestContextAdminSecondOrga);
       await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
         {
@@ -2160,10 +2286,12 @@ describe('deployment app', () => {
         user: contextUserWithPersonalOrga.user,
       });
 
+      // When loading the platform trial status for the personal space
       const platformTrialStatus = await DeploymentApp.loadPlatformTrialStatus(
         TEST_ORGANIZATIONS.SECOND_ORGANIZATION.USERS.SIMPLE.PERSONAL_SPACE_ID
       );
 
+      // Then hub_status and end_date are null
       expect(platformTrialStatus).toEqual({
         ongoingStandaloneTrials: [],
         isBlacklisted: false,
@@ -2173,11 +2301,15 @@ describe('deployment app', () => {
     });
 
     it('should throw if user does not belong in the organization', async () => {
+      // Given a user belonging to a different organization
       requestContext.set(requestContextAdminSecondOrga);
+
+      // When loading the platform trial status for an organization the user does not belong to
       const call = DeploymentApp.loadPlatformTrialStatus(
         TEST_ORGANIZATIONS.FILIGRAN.ID
       );
 
+      // Then it throws UserIsNotInOrganization
       await expect(call).rejects.toThrow(ErrorCode.UserIsNotInOrganization);
     });
 
@@ -2204,20 +2336,24 @@ describe('deployment app', () => {
 
   describe('reorderDeploymentRequestInQueue', () => {
     it('should throw when deployment request is not found', async () => {
+      // Given no matching deployment request
       vi.spyOn(
         DeploymentRequestDomain,
         'loadDeploymentRequestBy'
       ).mockResolvedValue(undefined);
 
+      // When reordering a deployment request to the top of the queue
       const call = DeploymentApp.reorderDeploymentRequestInQueue({
         id: uuidv4() as DeploymentRequestId,
         direction: ReorderDeploymentRequestInQueueDirection.Top,
       });
 
+      // Then it throws DeploymentRequestNotFound
       await expect(call).rejects.toThrow(ErrorCode.DeploymentRequestNotFound);
     });
 
     it('should throw when deployment request is not in queue', async () => {
+      // Given a deployment request that is not queued
       const deploymentRequest = {
         id: uuidv4() as DeploymentRequestId,
         hub_status: DeploymentRequestHubStatus.Active,
@@ -2229,17 +2365,20 @@ describe('deployment app', () => {
         'loadDeploymentRequestBy'
       ).mockResolvedValue(deploymentRequest);
 
+      // When reordering that deployment request to the top of the queue
       const call = DeploymentApp.reorderDeploymentRequestInQueue({
         id: deploymentRequest!.id,
         direction: ReorderDeploymentRequestInQueueDirection.Top,
       });
 
+      // Then it throws DeploymentRequestHubStatusNotQueued
       await expect(call).rejects.toThrow(
         ErrorCode.DeploymentRequestHubStatusNotQueued
       );
     });
 
     it('should reorder deployment request to top when direction is top', async () => {
+      // Given a queued deployment request
       const deploymentRequest = {
         id: uuidv4() as DeploymentRequestId,
         hub_status: DeploymentRequestHubStatus.Queued,
@@ -2255,11 +2394,13 @@ describe('deployment app', () => {
         .spyOn(DeploymentRequestDomain, 'reorderDeploymentRequestToTop')
         .mockResolvedValue();
 
+      // When reordering with direction Top
       const result = await DeploymentApp.reorderDeploymentRequestInQueue({
         id: uuidv4() as DeploymentRequestId,
         direction: ReorderDeploymentRequestInQueueDirection.Top,
       });
 
+      // Then the operation succeeds and the domain reorder-to-top function is called
       expect(result.success).toBeTruthy();
       expect(reorderDeploymentRequestToTopSpy).toHaveBeenCalledWith(
         deploymentRequest
@@ -2267,6 +2408,7 @@ describe('deployment app', () => {
     });
 
     it('should reorder deployment request up when direction is up', async () => {
+      // Given a queued deployment request
       const deploymentRequest = {
         id: uuidv4() as DeploymentRequestId,
         hub_status: DeploymentRequestHubStatus.Queued,
@@ -2282,11 +2424,13 @@ describe('deployment app', () => {
         .spyOn(DeploymentRequestDomain, 'reorderDeploymentRequestUp')
         .mockResolvedValue();
 
+      // When reordering with direction Up
       const result = await DeploymentApp.reorderDeploymentRequestInQueue({
         id: uuidv4() as DeploymentRequestId,
         direction: ReorderDeploymentRequestInQueueDirection.Up,
       });
 
+      // Then the operation succeeds and the domain reorder-up function is called
       expect(result.success).toBeTruthy();
       expect(reorderDeploymentRequestUpSpy).toHaveBeenCalledWith(
         deploymentRequest
@@ -2320,6 +2464,7 @@ describe('deployment app', () => {
         counts_in_orga_quota,
         target_state,
       }) => {
+        // Given a deployment request in the provided hub/actual state
         const initialDeployment =
           (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -2328,12 +2473,15 @@ describe('deployment app', () => {
             }
           )) as DeploymentRequest;
         const cancellationReason = isAdmin ? undefined : 'my reason';
+
+        // When cancelling the deployment request
         const deployment = await DeploymentApp.cancelDeploymentRequest(
           initialDeployment.id,
           isAdmin,
           cancellationReason
         );
 
+        // Then the deployment request is updated accordingly
         expect(deployment).toMatchObject({
           hub_status: DeploymentRequestHubStatus.Cancelled,
           target_state: target_state,
@@ -2370,57 +2518,71 @@ describe('deployment app', () => {
       }
     );
     it('should throw if deployment request does not exist', async () => {
+      // Given a deployment request id that does not exist
+      // When cancelling the deployment request
       const call = DeploymentApp.cancelDeploymentRequest(
         uuidv4() as DeploymentRequestId,
         false
       );
+
+      // Then it throws a not found error
       await expect(call).rejects.toThrow(
         NotFoundErrorCode.DeploymentRequestNotFound
       );
     });
 
     it('should throw if user is not in organization and not isAdmin', async () => {
+      // Given a deployment request owned by another organization
       const deployment =
         (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {}
         )) as DeploymentRequest;
-
       requestContext.set(requestContextAdminSecondOrga);
+
+      // When cancelling the deployment request as a non-admin
       const call = DeploymentApp.cancelDeploymentRequest(deployment.id, false);
+
+      // Then it throws a forbidden error
       await expect(call).rejects.toThrow(
         ForbiddenErrorCode.UserIsNotInOrganization
       );
     });
 
     it('should not throw if user is not in organization and isAdmin', async () => {
+      // Given a deployment request owned by another organization
       const deployment =
         (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {}
         )) as DeploymentRequest;
-
       requestContext.set(requestContextAdminSecondOrga);
+
+      // When cancelling the deployment request as an admin
       const response = await DeploymentApp.cancelDeploymentRequest(
         deployment.id,
         true
       );
+
+      // Then it succeeds
       expect(response).toBeTruthy();
     });
     it('should send a telemetry event', async () => {
+      // Given a deployment request and a fixed system time
       const deployment =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {}
         );
-
       vi.useFakeTimers();
       const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
       vi.setSystemTime(date);
 
+      // When cancelling the deployment request
       await DeploymentApp.cancelDeploymentRequest(
         deployment.id,
         false,
         'CancellationReason'
       );
 
+      // Then a telemetry event is sent
       expect(telemetrySpy).toHaveBeenCalledExactlyOnceWith({
         '@timestamp': '2025-02-03T13:12:15.000Z',
         event_type: TelemetryEventType.UPDATE_DEPLOYMENT,
@@ -2446,6 +2608,7 @@ describe('deployment app', () => {
     `(
       'should refuse to cancel a single product of a bundle $description',
       async ({ isAdmin }) => {
+        // Given a bundle with an active child product
         const bundle =
           await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -2465,16 +2628,17 @@ describe('deployment app', () => {
             }
           );
 
+        // When cancelling the child product directly
         const call = DeploymentApp.cancelDeploymentRequest(
           child.id,
           isAdmin,
           'my reason'
         );
 
+        // Then it throws a forbidden error and leaves both untouched
         await expect(call).rejects.toThrow(
           ForbiddenErrorCode.CantCancelBundleProduct
         );
-
         await TestHelper.deploymentRequest.assertProperties(child.id, {
           hub_status: DeploymentRequestHubStatus.Active,
           cancellation_date: null,
@@ -2488,6 +2652,7 @@ describe('deployment app', () => {
       }
     );
     it('should send a mail to the trial requester', async () => {
+      // Given a deployment request requested by a user in another organization
       const deployment =
         (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {
@@ -2496,8 +2661,10 @@ describe('deployment app', () => {
           }
         )) as DeploymentRequest;
 
+      // When cancelling the deployment request as an admin
       await DeploymentApp.cancelDeploymentRequest(deployment.id, true);
 
+      // Then a cancellation email is sent to the trial requester
       expect(mockSendMail).toHaveBeenCalledWith({
         to: TEST_ORGANIZATIONS.SECOND_ORGANIZATION.USERS.ADMIN_ORGA.EMAIL,
         template: 'free_trial_cancelled',
@@ -2515,12 +2682,12 @@ describe('deployment app', () => {
     `(
       'should log a $expectedLevel when deleting a missing audience (404) for $platformIdentifier',
       async ({ platformIdentifier, expectedLevel }) => {
+        // Given a deletion of the Auth0 audience that fails with a 404
         vi.spyOn(auth0ClientMock, 'deleteAudienceAPI').mockRejectedValue({
           statusCode: 404,
         });
         const warnSpy = vi.spyOn(logApp, 'warn').mockImplementation(() => {});
         const errorSpy = vi.spyOn(logApp, 'error').mockImplementation(() => {});
-
         const deployment =
           (await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -2531,8 +2698,10 @@ describe('deployment app', () => {
             }
           )) as DeploymentRequest;
 
+        // When cancelling the deployment request
         await DeploymentApp.cancelDeploymentRequest(deployment.id, true);
 
+        // Then it logs at the expected level depending on the platform
         if (expectedLevel === 'warn') {
           expect(warnSpy).toHaveBeenCalledWith(
             'No Auth0 audience to delete for OpenAEV trial',
@@ -2587,12 +2756,14 @@ describe('deployment app', () => {
       });
 
       it('should cancel the bundle and all its children with the cancellation reason', async () => {
+        // When cancelling the bundle
         await DeploymentApp.cancelDeploymentRequest(
           bundle.id,
           false,
           'my reason'
         );
 
+        // Then the bundle and all its children are cancelled with the reason
         for (const { id } of [bundle, childOpencti, childXtmone]) {
           const cancelled =
             await DeploymentRequestDomain.loadDeploymentRequestBy({ id });
@@ -2607,12 +2778,14 @@ describe('deployment app', () => {
       });
 
       it('should send one telemetry event per deployment request', async () => {
+        // When cancelling the bundle
         await DeploymentApp.cancelDeploymentRequest(
           bundle.id,
           false,
           'my reason'
         );
 
+        // Then one telemetry event is sent per deployment request
         expect(telemetrySpy).toHaveBeenCalledTimes(3);
         for (const { id } of [bundle, childOpencti, childXtmone]) {
           expect(telemetrySpy).toHaveBeenCalledWith(
@@ -2627,6 +2800,7 @@ describe('deployment app', () => {
       });
 
       it('should leave the whole bundle untouched when one row fails to be cancelled', async () => {
+        // Given the second update call failing
         const originalUpdate =
           DeploymentRequestDomain.updateDeploymentRequestById;
         let callCount = 0;
@@ -2641,14 +2815,15 @@ describe('deployment app', () => {
           return originalUpdate(id, data);
         });
 
+        // When cancelling the bundle
         const call = DeploymentApp.cancelDeploymentRequest(
           bundle.id,
           false,
           'my reason'
         );
 
+        // Then it throws and leaves the whole bundle untouched
         await expect(call).rejects.toThrow('boom');
-
         for (const { id, hub_status } of [bundle, childOpencti, childXtmone]) {
           const untouched =
             await DeploymentRequestDomain.loadDeploymentRequestBy({ id });
@@ -2719,6 +2894,7 @@ describe('deployment app', () => {
 
     describe('increase capacity', () => {
       it('should move queued request to pending when there is space available', async () => {
+        // Given a quota with active, pending and queued requests
         await initQuota({ capacity: 2, availability: 0 });
         const { id: activeRequestId } = await insertRequest(
           DeploymentRequestHubStatus.Active
@@ -2736,12 +2912,14 @@ describe('deployment app', () => {
           4
         );
 
+        // When increasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 4,
         });
 
+        // Then the quota is updated and the queued requests become pending
         await assertQuota({ capacity: 4, availability: 0 });
 
         await TestHelper.deploymentRequest.assertProperties(activeRequestId, {
@@ -2762,6 +2940,7 @@ describe('deployment app', () => {
       });
 
       it('should not move queued request to pending when there is just enough space for active', async () => {
+        // Given a quota with just enough space for the active requests
         await initQuota({ capacity: 1, availability: -1 });
         const { id: activeRequestId1 } = await insertRequest(
           DeploymentRequestHubStatus.Active
@@ -2773,12 +2952,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Queued
         );
 
+        // When increasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 2,
         });
 
+        // Then the quota is updated and the queued request stays queued
         await assertQuota({ capacity: 2, availability: 0 });
 
         await TestHelper.deploymentRequest.assertProperties(activeRequestId1, {
@@ -2793,6 +2974,7 @@ describe('deployment app', () => {
       });
 
       it('should not move queued request to pending when there is not more space available', async () => {
+        // Given a quota with no additional space for the queued request
         await initQuota({ capacity: 1, availability: -2 });
         const { id: activeRequestId1 } = await insertRequest(
           DeploymentRequestHubStatus.Active
@@ -2807,12 +2989,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Queued
         );
 
+        // When increasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 2,
         });
 
+        // Then the quota is updated and the queued request stays queued
         await assertQuota({ capacity: 2, availability: -1 });
 
         await TestHelper.deploymentRequest.assertProperties(activeRequestId1, {
@@ -2830,6 +3014,7 @@ describe('deployment app', () => {
       });
 
       it('should send telemetry event for each request moved', async () => {
+        // Given a quota with queued requests and a fixed system time
         vi.useFakeTimers();
         const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
         vi.setSystemTime(date);
@@ -2843,12 +3028,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Queued
         );
 
+        // When increasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 2,
         });
 
+        // Then the quota is updated and a telemetry event is sent for each moved request
         await assertQuota({ capacity: 2, availability: 0 });
 
         expect(telemetrySpy).toHaveBeenCalledTimes(2);
@@ -2888,17 +3075,20 @@ describe('deployment app', () => {
     });
     describe('decrease capacity', () => {
       it('should release pending requests from availability', async () => {
+        // Given a quota with a pending request
         await initQuota({ capacity: 1, availability: 0 });
         const { id: pendingRequestId } = await insertRequest(
           DeploymentRequestHubStatus.Pending
         );
 
+        // When decreasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 0,
         });
 
+        // Then the quota is updated and the pending request becomes queued
         await assertQuota({ capacity: 0, availability: 0 });
 
         await TestHelper.deploymentRequest.assertProperties(pendingRequestId, {
@@ -2907,6 +3097,7 @@ describe('deployment app', () => {
       });
 
       it('should only release pending requests until availability is equal to zero', async () => {
+        // Given a quota with two pending requests
         await initQuota({ capacity: 2, availability: 0 });
         const { id: pendingRequestId1 } = await insertRequest(
           DeploymentRequestHubStatus.Pending
@@ -2915,12 +3106,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Pending
         );
 
+        // When decreasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 1,
         });
 
+        // Then the quota is updated and only one pending request becomes queued
         await assertQuota({ capacity: 1, availability: 0 });
 
         await TestHelper.deploymentRequest.assertProperties(pendingRequestId1, {
@@ -2932,6 +3125,7 @@ describe('deployment app', () => {
       });
 
       it('should move pending requests to queued with the right ordering', async () => {
+        // Given a quota with pending and queued requests in various orderings
         await initQuota({ capacity: 2, availability: 0 });
         const { id: pendingRequestId1 } = await insertRequest(
           DeploymentRequestHubStatus.Pending,
@@ -2950,12 +3144,14 @@ describe('deployment app', () => {
           5
         );
 
+        // When decreasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 0,
         });
 
+        // Then the quota is updated and pending requests are queued with the right ordering
         await assertQuota({ capacity: 0, availability: 0 });
 
         await TestHelper.deploymentRequest.assertProperties(pendingRequestId1, {
@@ -2977,6 +3173,7 @@ describe('deployment app', () => {
       });
 
       it('should move only pending requests to free place when new availability is negative', async () => {
+        // Given a quota with active, pending and queued requests
         await initQuota({ capacity: 3, availability: 0 });
         const { id: activeRequestId1 } = await insertRequest(
           DeploymentRequestHubStatus.Active
@@ -2991,12 +3188,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Queued
         );
 
+        // When decreasing the quota capacity to a negative availability
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 1,
         });
 
+        // Then the quota is updated and only the pending request is moved to queued
         await assertQuota({ capacity: 1, availability: -1 });
 
         await TestHelper.deploymentRequest.assertProperties(activeRequestId1, {
@@ -3013,6 +3212,7 @@ describe('deployment app', () => {
         });
       });
       it('should not move pending requests to queue when new availability is equal to zero', async () => {
+        // Given a quota with an active and a pending request
         await initQuota({ capacity: 3, availability: 1 });
         const { id: activeRequestId } = await insertRequest(
           DeploymentRequestHubStatus.Active
@@ -3021,12 +3221,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Pending
         );
 
+        // When decreasing the quota capacity so availability reaches zero
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 2,
         });
 
+        // Then the quota is updated and the pending request stays pending
         await assertQuota({ capacity: 2, availability: 0 });
 
         await TestHelper.deploymentRequest.assertProperties(activeRequestId, {
@@ -3037,6 +3239,7 @@ describe('deployment app', () => {
         });
       });
       it('should not move pending requests to queue when new availability is positive', async () => {
+        // Given a quota with an active and a pending request
         await initQuota({ capacity: 4, availability: 2 });
         const { id: activeRequestId } = await insertRequest(
           DeploymentRequestHubStatus.Active
@@ -3044,12 +3247,15 @@ describe('deployment app', () => {
         const { id: pendingRequestId } = await insertRequest(
           DeploymentRequestHubStatus.Pending
         );
+
+        // When decreasing the quota capacity while availability remains positive
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 3,
         });
 
+        // Then the quota is updated and the pending request stays pending
         await assertQuota({ capacity: 3, availability: 1 });
 
         await TestHelper.deploymentRequest.assertProperties(activeRequestId, {
@@ -3061,6 +3267,7 @@ describe('deployment app', () => {
       });
 
       it('should send telemetry event for each request moved', async () => {
+        // Given a quota with pending requests and a fixed system time
         vi.useFakeTimers();
         const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
         vi.setSystemTime(date);
@@ -3074,12 +3281,14 @@ describe('deployment app', () => {
           DeploymentRequestHubStatus.Pending
         );
 
+        // When decreasing the quota capacity
         await DeploymentApp.updateDeploymentQuotaCapacity({
           platformIdentifier,
           region,
           newCapacity: 0,
         });
 
+        // Then the quota is updated and a telemetry event is sent for each moved request
         await assertQuota({ capacity: 0, availability: 0 });
 
         expect(telemetrySpy).toHaveBeenCalledTimes(2);
@@ -3128,6 +3337,7 @@ describe('deployment app', () => {
     });
 
     it('should expire past trials only', async () => {
+      // Given an expired trial and a still-running trial
       vi.useFakeTimers();
       const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
       vi.setSystemTime(date);
@@ -3151,8 +3361,10 @@ describe('deployment app', () => {
           }
         );
 
+      // When expiring trials
       await DeploymentApp.expireTrials();
 
+      // Then the expired trial is marked expired and the non-expired trial is untouched
       const expiredDeploymentRequest =
         await DeploymentRequestDomain.loadDeploymentRequestBy({
           id: expiredTrial?.id as DeploymentRequestId,
@@ -3185,6 +3397,7 @@ describe('deployment app', () => {
     `(
       `should not expire trials in status $hub_status`,
       async ({ hub_status, target_state }) => {
+        // Given a trial in a status that should not be expired
         vi.useFakeTimers();
         const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
         vi.setSystemTime(date);
@@ -3198,8 +3411,10 @@ describe('deployment app', () => {
             }
           );
 
+        // When expiring trials
         await DeploymentApp.expireTrials();
 
+        // Then the trial is left untouched
         const expiredDeploymentRequest =
           await DeploymentRequestDomain.loadDeploymentRequestBy({
             id: trial?.id as DeploymentRequestId,
@@ -3212,6 +3427,7 @@ describe('deployment app', () => {
       }
     );
     it('should send a mail to the requester', async () => {
+      // Given an expired trial belonging to a requester
       vi.useFakeTimers();
       const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
       vi.setSystemTime(date);
@@ -3227,8 +3443,10 @@ describe('deployment app', () => {
         }
       );
 
+      // When expiring trials
       await DeploymentApp.expireTrials();
 
+      // Then a mail is sent to the requester
       expect(mockSendMail).toHaveBeenCalledWith({
         to: TEST_ORGANIZATIONS.SECOND_ORGANIZATION.USERS.ADMIN_ORGA.EMAIL,
         template: 'free_trial_expired',
@@ -3240,6 +3458,7 @@ describe('deployment app', () => {
     });
 
     it('should send a telemetry event', async () => {
+      // Given an expired trial with organization/requester information
       vi.useFakeTimers();
       const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
       vi.setSystemTime(date);
@@ -3260,8 +3479,10 @@ describe('deployment app', () => {
           }
         );
 
+      // When expiring trials
       await DeploymentApp.expireTrials();
 
+      // Then a telemetry event is emitted with the expiry details
       expect(telemetrySpy).toHaveBeenCalledExactlyOnceWith({
         '@timestamp': '2025-02-03T13:12:15.000Z',
         event_type: TelemetryEventType.UPDATE_DEPLOYMENT,
@@ -3321,8 +3542,10 @@ describe('deployment app', () => {
       });
 
       it('should expire the bundle and all its children', async () => {
+        // When expiring trials
         await DeploymentApp.expireTrials();
 
+        // Then the bundle and all its children are expired
         for (const { id } of [bundle, childOpencti, childXtmone]) {
           await TestHelper.deploymentRequest.assertProperties(id, {
             hub_status: DeploymentRequestHubStatus.Expired,
@@ -3332,6 +3555,7 @@ describe('deployment app', () => {
       });
 
       it('should not expire a standalone trial that is still running', async () => {
+        // Given a standalone trial that has not reached its end date
         const ongoingStandalone =
           await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
             {
@@ -3341,8 +3565,10 @@ describe('deployment app', () => {
             }
           );
 
+        // When expiring trials
         await DeploymentApp.expireTrials();
 
+        // Then the trial remains active
         await TestHelper.deploymentRequest.assertProperties(
           ongoingStandalone.id,
           { hub_status: DeploymentRequestHubStatus.Active }
@@ -3350,6 +3576,7 @@ describe('deployment app', () => {
       });
 
       it('should leave the whole bundle untouched when one row fails to be expired', async () => {
+        // Given an update that fails for the second call
         const originalUpdate =
           DeploymentRequestDomain.updateDeploymentRequestById;
         let callCount = 0;
@@ -3364,8 +3591,10 @@ describe('deployment app', () => {
           return originalUpdate(id, data);
         });
 
+        // When expiring trials
         await DeploymentApp.expireTrials();
 
+        // Then the whole bundle is left untouched
         for (const { id } of [bundle, childOpencti, childXtmone]) {
           const untouched =
             await DeploymentRequestDomain.loadDeploymentRequestBy({ id });
@@ -3374,6 +3603,7 @@ describe('deployment app', () => {
       });
 
       it('should not re-expire a child that is already cancelled', async () => {
+        // Given a bundle child that is already cancelled
         await DeploymentRequestDomain.updateDeploymentRequestById(
           childXtmone.id,
           {
@@ -3382,8 +3612,10 @@ describe('deployment app', () => {
           }
         );
 
+        // When expiring trials
         await DeploymentApp.expireTrials();
 
+        // Then the cancelled child stays cancelled while the bundle is expired
         await TestHelper.deploymentRequest.assertProperties(childXtmone.id, {
           hub_status: DeploymentRequestHubStatus.Cancelled,
         });
@@ -3412,6 +3644,7 @@ describe('deployment app', () => {
       `(
         'should not free place when request hub status is $hub_status',
         async ({ hub_status }) => {
+          // Given a deployment request in a status not counted in quotas
           const deploymentRequest =
             await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
               {
@@ -3419,18 +3652,21 @@ describe('deployment app', () => {
               }
             );
 
+          // When releasing its place
           await DeploymentApp.releaseDeploymentRequestPlace(
             deploymentRequest!.hub_status,
             deploymentRequest!.platform_identifier,
             deploymentRequest!.region
           );
 
+          // Then no place is freed
           expect(freePlaceSpy).not.toHaveBeenCalled();
         }
       );
     });
 
     it('should set one queued request as pending and not free place', async () => {
+      // Given an active deployment request and a queued one waiting to take its place
       const deploymentRequestToRelease =
         await TestHelper.deploymentRequest.createWithServiceInstanceAndSubscription(
           {
@@ -3445,12 +3681,14 @@ describe('deployment app', () => {
         .spyOn(DeploymentRequestDomain, 'setFirstQueuedRequestAsPending')
         .mockResolvedValue(queuedDeploymentRequest);
 
+      // When releasing the place of the active deployment request
       await DeploymentApp.releaseDeploymentRequestPlace(
         deploymentRequestToRelease!.hub_status,
         deploymentRequestToRelease!.platform_identifier,
         deploymentRequestToRelease!.region
       );
 
+      // Then the queued request is set as pending and no place is freed
       expect(setFirstQueuedRequestAsPendingSpy).toHaveBeenCalledWith(
         deploymentRequestToRelease!.platform_identifier,
         deploymentRequestToRelease!.region
@@ -3459,6 +3697,7 @@ describe('deployment app', () => {
     });
 
     it('should free place when deployment request was not moved to pending', async () => {
+      // Given no queued request available to take the released place
       vi.spyOn(
         DeploymentRequestDomain,
         'setFirstQueuedRequestAsPending'
@@ -3471,12 +3710,14 @@ describe('deployment app', () => {
           }
         );
 
+      // When releasing the deployment request's place
       await DeploymentApp.releaseDeploymentRequestPlace(
         deploymentRequest!.hub_status,
         deploymentRequest!.platform_identifier,
         deploymentRequest!.region
       );
 
+      // Then the place is freed
       expect(freePlaceSpy).toHaveBeenCalledWith(
         deploymentRequest!.platform_identifier,
         deploymentRequest!.region
@@ -3485,6 +3726,7 @@ describe('deployment app', () => {
 
     describe('telemetry', () => {
       it('should not send telemetry event when deployment request was not moved to pending', async () => {
+        // Given no queued request available to take the released place
         vi.spyOn(
           DeploymentRequestDomain,
           'setFirstQueuedRequestAsPending'
@@ -3497,16 +3739,19 @@ describe('deployment app', () => {
             }
           );
 
+        // When releasing the deployment request's place
         await DeploymentApp.releaseDeploymentRequestPlace(
           deploymentRequest!.hub_status,
           deploymentRequest!.platform_identifier,
           deploymentRequest!.region
         );
 
+        // Then no telemetry event is sent
         expect(telemetrySpy).not.toHaveBeenCalled();
       });
 
       it('should send telemetry event when deployment request was moved to pending', async () => {
+        // Given a queued deployment request that will be moved to pending
         vi.useFakeTimers();
         const date = new Date(Date.UTC(2025, 1, 3, 13, 12, 15));
         vi.setSystemTime(date);
@@ -3536,12 +3781,14 @@ describe('deployment app', () => {
             }
           );
 
+        // When releasing the active deployment request's place
         await DeploymentApp.releaseDeploymentRequestPlace(
           deploymentRequest!.hub_status,
           deploymentRequest!.platform_identifier,
           deploymentRequest!.region
         );
 
+        // Then a telemetry event is emitted for the promoted queued request
         expect(telemetrySpy).toHaveBeenCalledExactlyOnceWith({
           '@timestamp': '2025-02-03T13:12:15.000Z',
           event_type: TelemetryEventType.UPDATE_DEPLOYMENT,
