+++ apps/backend/src/modules/deployment/deployment.app.ts
@@ -46,9 +46,7 @@ import {
 } from '../../portal.const';
 import { securityGuard } from '../../security/guard';
 import { sendMail } from '../../server/mail-service';
-import { auth0Client } from '../../thirdparty/auth0/client';
 import { logApp } from '../../utils/app-logger.util';
-import { getErrorNumberProperty } from '../../utils/error/error-guard.util';
 import {
   BadRequestErrorCode,
   ErrorCode,
@@ -79,8 +77,8 @@ import { DeploymentHelper } from './deployment.helper';
 import { DeploymentQuotaDomain } from './quota/deployment.quota.domain';
 
 export const XTM_PLATFORM_BUNDLE_SERVICE_INSTANCE_NAME = 'XTM Platform Bundle';
-export const BUNDLE_ACTIVATION_CANCELLATION_REASON =
-  'Cancelled automatically when the XTM Platform trial became active';
+export const BUNDLE_REQUEST_CANCELLATION_REASON =
+  'Other: Cancelled automatically when the XTM Platform trial was requested';
 
 export const DeploymentApp = {
   createDeploymentRequest: async (
@@ -398,12 +396,13 @@ export const DeploymentApp = {
       const updated: DeploymentRequestModel[] = [];
 
       for (const request of family) {
-        const updatedRequest = await applyCancellationToDeploymentRequest({
-          deploymentRequest: request,
-          userId: user.id,
-          isAdmin,
-          cancellationReason,
-        });
+        const updatedRequest =
+          await DeploymentRequestDomain.applyCancellationToDeploymentRequest({
+            deploymentRequest: request,
+            userId: user.id,
+            isAdmin,
+            cancellationReason,
+          });
         updated.push(updatedRequest);
 
         await sendDeploymentRequestCancelledNotifications(
@@ -416,7 +415,9 @@ export const DeploymentApp = {
     });
 
     for (const previousRequest of family) {
-      await deleteDeploymentRequestAudience(previousRequest);
+      await DeploymentRequestDomain.deleteDeploymentRequestAudience(
+        previousRequest
+      );
     }
 
     const [updatedDeploymentRequest] = updatedFamily;
@@ -458,7 +459,7 @@ export const DeploymentApp = {
       }
 
       for (const request of family) {
-        await deleteDeploymentRequestAudience(request);
+        await DeploymentRequestDomain.deleteDeploymentRequestAudience(request);
       }
     }
   },
@@ -601,41 +602,6 @@ type ValidatedDeploymentRequestProducts =
       platformIdentifier: PlatformIdentifier;
     };
 
-const validateUseCasesByProduct = (
-  input: CreateDeploymentRequestInput,
-  uniqueProducts: PlatformIdentifier[]
-): void => {
-  const productsRequiringUseCase = uniqueProducts.filter(
-    (platformIdentifier) => platformIdentifier !== PlatformIdentifier.Xtmone
-  );
-  const everyProductHasUseCase = productsRequiringUseCase.every(
-    (platformIdentifier) =>
-      input.use_cases_by_product?.some(
-        (entry) => entry.platform_identifier === platformIdentifier
-      )
-  );
-  const everyEntryTargetsAValidProduct =
-    input.use_cases_by_product?.every(
-      (entry) =>
-        entry.platform_identifier !== PlatformIdentifier.Xtmone &&
-        uniqueProducts.includes(entry.platform_identifier)
-    ) ?? true;
-  const targetedProducts =
-    input.use_cases_by_product?.map((entry) => entry.platform_identifier) ?? [];
-  const hasDuplicatedEntry =
-    targetedProducts.length !== new Set(targetedProducts).size;
-
-  if (!everyEntryTargetsAValidProduct) {
-    throw new Error(BadRequestErrorCode.InvalidUseCasesForProducts);
-  }
-  if (!everyProductHasUseCase) {
-    throw new Error(BadRequestErrorCode.InvalidUseCasesForProducts);
-  }
-  if (hasDuplicatedEntry) {
-    throw new Error(BadRequestErrorCode.InvalidUseCasesForProducts);
-  }
-};
-
 const validateDeploymentRequestProducts = (
   input: CreateDeploymentRequestInput
 ): ValidatedDeploymentRequestProducts => {
@@ -647,7 +613,7 @@ const validateDeploymentRequestProducts = (
     if (!includesXtmone || !hasAtLeastOneOtherProduct) {
       throw new Error(BadRequestErrorCode.InvalidProductsForDeploymentType);
     }
-    validateUseCasesByProduct(input, uniqueProducts);
+    DeploymentHelper.validateUseCasesByProduct(input, uniqueProducts);
     return {
       type: DeploymentRequestDeploymentType.Bundle,
       products: uniqueProducts,
@@ -660,7 +626,7 @@ const validateDeploymentRequestProducts = (
   if (!hasSingleProduct || isXtmone) {
     throw new Error(BadRequestErrorCode.InvalidProductsForDeploymentType);
   }
-  validateUseCasesByProduct(input, uniqueProducts);
+  DeploymentHelper.validateUseCasesByProduct(input, uniqueProducts);
   return { type: DeploymentRequestDeploymentType.Trial, platformIdentifier };
 };
 
@@ -910,6 +876,10 @@ const createBundleDeploymentRequest = async ({
         parent_id: null,
       });
 
+    await DeploymentRequestDomain.cancelOngoingStandaloneTrialsForBundle(
+      bundleDeploymentRequest
+    );
+
     try {
       const createDeploymentEvent = TelemetryHelper.buildCreateDeploymentEvent(
         chosenOrganization,
@@ -972,96 +942,6 @@ const loadDeploymentRequestForUpdate = async (
   return deploymentRequest;
 };
 
-const logDeleteAudienceError = (
-  error: unknown,
-  deploymentRequest: Pick<DeploymentRequestModel, 'id' | 'platform_identifier'>
-) => {
-  const isExpectedOpenaevAudienceMiss =
-    getErrorNumberProperty(error, 'statusCode') === 404 &&
-    deploymentRequest.platform_identifier === PlatformIdentifier.Openaev;
-
-  if (isExpectedOpenaevAudienceMiss) {
-    logApp.warn('No Auth0 audience to delete for OpenAEV trial', {
-      deploymentRequestId: deploymentRequest.id,
-    });
-    return;
-  }
-
-  logApp.error('Unable to delete audience', {
-    error,
-    deploymentRequestId: deploymentRequest.id,
-  });
-};
-
-const applyCancellationToDeploymentRequest = async ({
-  deploymentRequest,
-  userId,
-  isAdmin,
-  cancellationReason,
-}: {
-  deploymentRequest: DeploymentRequestModel;
-  userId: UserId;
-  isAdmin: boolean;
-  cancellationReason?: string;
-}): Promise<DeploymentRequestModel> => {
-  const previousHubStatus = deploymentRequest.hub_status;
-
-  const countsInOrgaQuota =
-    isAdmin ||
-    ![
-      DeploymentRequestPlatformState.Unprovisioned,
-      DeploymentRequestPlatformState.Provisioning,
-    ].includes(
-      deploymentRequest.actual_state ??
-        DeploymentRequestPlatformState.Unprovisioned
-    );
-
-  const target_state =
-    deploymentRequest.actual_state ===
-    DeploymentRequestPlatformState.Unprovisioned
-      ? DeploymentRequestPlatformState.Unprovisioned
-      : DeploymentRequestPlatformState.Removed;
-
-  return DeploymentQuotaDomain.withLockedQuotaTransaction(
-    {
-      platformIdentifier: deploymentRequest.platform_identifier,
-      region: deploymentRequest.region,
-    },
-    async () => {
-      const updatedDeploymentRequest =
-        await DeploymentRequestDomain.updateDeploymentRequestById(
-          deploymentRequest.id,
-          {
-            hub_status: DeploymentRequestHubStatus.Cancelled,
-            target_state: target_state,
-            cancellation_date: new Date(),
-            cancellation_user_id: userId,
-            cancellation_reason: cancellationReason,
-            counts_in_orga_quota: countsInOrgaQuota,
-          }
-        );
-      if (!updatedDeploymentRequest) {
-        throw new Error(NotFoundErrorCode.DeploymentRequestNotFound);
-      }
-      if (!countsInOrgaQuota) {
-        await ServiceInstanceDomain.updateServiceInstance(
-          deploymentRequest.service_instance_id,
-          {
-            creation_status: ServiceInstanceCreationStatus.Disabled,
-          }
-        );
-      }
-      await DeploymentApp.releaseDeploymentRequestPlace(
-        previousHubStatus,
-        deploymentRequest.platform_identifier,
-        deploymentRequest.region
-      );
-
-      return updatedDeploymentRequest;
-    }
-  );
-};
-
 const sendDeploymentRequestCancelledNotifications = async (
   deploymentRequest: DeploymentRequestModel,
   userId: UserId
@@ -1091,26 +971,6 @@ const sendDeploymentRequestCancelledNotifications = async (
       error,
       deploymentRequestId: deploymentRequest.id,
     });
-  }
-};
-
-const deleteDeploymentRequestAudience = async (
-  deploymentRequest: DeploymentRequestModel
-): Promise<void> => {
-  try {
-    if (deploymentRequest.platform_id) {
-      await auth0Client.deleteAudienceAPI(
-        deploymentRequest.organization_requester_id,
-        deploymentRequest.platform_id
-      );
-    } else {
-      logApp.error('Unable to delete audience', {
-        error: 'missing platform_id',
-        deploymentRequestId: deploymentRequest.id,
-      });
-    }
-  } catch (error) {
-    logDeleteAudienceError(error, deploymentRequest);
   }
 };
 
@@ -1288,25 +1148,6 @@ const recomputeBundleDates = async (bundleId: DeploymentRequestId) => {
   );
 };
 
-const cancelOngoingStandaloneTrialsForBundle = async (
-  bundle: DeploymentRequestModel
-) => {
-  const standaloneTrials =
-    await DeploymentRequestDomain.loadOngoingStandaloneTrialsForOrganization(
-      bundle.organization_requester_id
-    );
-
-  for (const trial of standaloneTrials) {
-    await applyCancellationToDeploymentRequest({
-      deploymentRequest: trial,
-      userId: bundle.user_requester_id,
-      isAdmin: false,
-      cancellationReason: BUNDLE_ACTIVATION_CANCELLATION_REASON,
-    });
-    await deleteDeploymentRequestAudience(trial);
-  }
-};
-
 const applyDeploymentRequestUpdateInQuotaTransaction = async ({
   deploymentRequest,
   deploymentRequestId,
@@ -1334,13 +1175,6 @@ const applyDeploymentRequestUpdateInQuotaTransaction = async ({
             hub_status: newStatus,
           }
         );
-
-        const becomesActive =
-          newStatus === DeploymentRequestHubStatus.Active &&
-          deploymentRequest.hub_status !== DeploymentRequestHubStatus.Active;
-        if (becomesActive) {
-          await cancelOngoingStandaloneTrialsForBundle(deploymentRequest);
-        }
 
         return;
       }
