--- apps/frontend/graphql/trial/trial-deployments-eligibility.query.graphql
@@ -1,6 +1,0 @@
-query TrialDeploymentsEligibility($input: TrialDeploymentsInput!) {
-  trialDeployments(input: $input) {
-    availableTrials
-    isBlacklisted
-  }
-}
