+++ apps/frontend/graphql/trial/trial.mutation.graphql
