+++ apps/frontend/graphql/trial/trial.query.graphql
@@ -6,3 +6,10 @@ query PlatformTrialStatus($organizationId: OrganizationId!) {
     ongoingStandaloneTrials
   }
 }
+
+query TrialDeploymentsEligibility($input: TrialDeploymentsInput!) {
+  trialDeployments(input: $input) {
+    availableTrials
+    isBlacklisted
+  }
+}
